Refactoring ApiMgtDAO to OrganizationPurgeDAO
